+++ R/seurat.R
@@ -478,6 +478,242 @@ RenameAssays <- function(object, ...) {
   return(object)
 }
 
+#' Save Seurat Objects to RDS files
+#'
+#' @param object A \code{\link{Seurat}} object
+#' @param file Path to save \code{object} to; defaults to
+#' \code{file.path(getwd(), paste0(Project(object), ".Rds"))}
+#' @param destdir Destination directory for on-disk layers saved in
+#' \dQuote{\code{\Sexpr[stage=render]{tempdir()}}}
+#' @param relative Save relative paths instead of absolute ones
+#' @inheritDotParams base::saveRDS
+#'
+#' @return Invisibly returns \code{file}
+#'
+#' @export
+#'
+#' @examples
+#' if (requireNamespace("HDF5Array") && requireNamespace("fs")) {
+#'   out <- tempfile(fileext = ".Rds")
+#'   pbmc_small[["disk"]] <- CreateAssay5Object(list(
+#'     mem = LayerData(pbmc_small, "counts"),
+#'     disk = as(LayerData(pbmc_small, "counts"), "HDF5Array")
+#'   ))
+#'   SaveSeuratRds(pbmc_small, file = out)
+#'   obj <- readRDS(out)
+#'   Tool(obj, "SaveSeuratRds")
+#' }
+#'
+SaveSeuratRds <- function(
+  object,
+  file = NULL,
+  destdir = NULL,
+  relative = FALSE,
+  ...
+) {
+  file <- file %||% file.path(getwd(), paste0(Project(object = object), '.Rds'))
+  file <- normalizePath(path = file, mustWork = FALSE)
+  # Cache v5 assays
+  assays <- .FilterObjects(object = object, classes.keep = 'StdAssay')
+  p <- progressor(along = assays, auto_finish = TRUE)
+  on.exit(expr = p(type = 'finish'), add = TRUE)
+  p(
+    message = paste(
+      "Looking for on-disk matrices in",
+      length(x = assays),
+      "assays"
+    ),
+    class = 'sticky',
+    amount = 0
+  )
+  cache <- vector(mode = 'list', length = length(x = assays))
+  names(x = cache) <- assays
+  tdir <- normalizePath(path = tempdir()) # because macOS is weird
+  destdir <- destdir %||% dirname(path = file)
+  if (!is_na(x = destdir) || isTRUE(x = relative)) {
+    check_installed(
+      pkg = 'fs',
+      reason = 'for moving on-disk matrices out of temp'
+    )
+  }
+  for (assay in assays) {
+    p(
+      message = paste("Searching through assay", assay),
+      class = 'sticky',
+      amount = 0
+    )
+    df <- lapply(
+      X = Layers(object = object[[assay]]),
+      FUN = function(lyr) {
+        ldat <- LayerData(object = object[[assay]], layer = lyr)
+        path <- .FilePath(x = ldat)
+        if (is.null(x = path)) {
+          return(NULL)
+        }
+        return(data.frame(
+          layer = lyr,
+          path = path,
+          class = paste(class(x = ldat), collapse = ','),
+          pkg = .ClassPkg(object = ldat),
+          fxn = .DiskLoad(x = ldat)
+        ))
+      }
+    )
+    df <- do.call(what = 'rbind', args = df)
+    if (is.null(x = df) || !nrow(x = df)) {
+      p(message = "No on-disk layers found", class = 'sticky', amount = 0)
+      next
+    }
+    if (!is_na(x = destdir)) {
+      for (i in seq_len(length.out = nrow(x = df))) {
+        pth <- df$path[i]
+        if (substr(x = pth, start = 1L, stop = nchar(x = tdir)) == tdir) {
+          p(
+            message = paste(
+              "Moving layer",
+              sQuote(x = df$layer[i], q = FALSE),
+              "out of temporary storage to",
+              sQuote(x = destdir, q = FALSE)
+            ),
+            class = 'sticky',
+            amount = 0
+          )
+          df[i, 'path'] <- as.character(x = fs::file_move(
+            path = pth,
+            new_path = destdir
+          ))
+        }
+      }
+    }
+    if (isTRUE(x = relative)) {
+      p(
+        message = paste(
+          "Adjusting paths to be relative to",
+          sQuote(x = dirname(path = file), q = FALSE)
+        ),
+        class = 'sticky',
+        amount = 0
+      )
+      df$path <- as.character(x = fs::path_rel(
+        path = df$path,
+        start = dirname(path = file)
+      ))
+    }
+    df$assay <- assay
+    cache[[assay]] <- df
+    if (nrow(x = df) == length(x = Layers(object = object[[assay]]))) {
+      p(
+        message = paste("Clearing layers from", assay),
+        class = 'sticky',
+        amount = 0
+      )
+      adata <- S4ToList(object = object[[assay]])
+      adata$layers <- list()
+      adata$default <- 0L
+      adata$cells <- LogMap(y = colnames(x = object[[assay]]))
+      adata$features <- LogMap(y = rownames(x = object[[assay]]))
+      object[[assay]] <- ListToS4(x = adata)
+    } else {
+      p(
+        message = paste("Clearing", nrow(x = df), "layers from", assay),
+        class = 'sticky',
+        amount = 0
+      )
+      for (layer in df$layer) {
+        LayerData(object = object[[assay]], layer = layer) <- NULL
+      }
+    }
+    p()
+  }
+  cache <- do.call(what = 'rbind', args = cache)
+  if (!is.null(x = cache) && nrow(x = cache)) {
+    p(message = "Saving on-disk cache to object", class = 'sticky', amount = 0)
+    row.names(x = cache) <- NULL
+    Tool(object = object) <- cache
+  }
+  saveRDS(object = object, file = file, ...)
+  return(invisible(x = file))
+}
+
+#' @export
+#'
+LoadSeuratRds <- function(file, ...) {
+  object <- readRDS(file = file, ...)
+  cache <- Tool(object = object, slot = 'SaveSeuratRds')
+  reqd.cols <- c('layer', 'path', 'class', 'pkg', 'fxn', 'assay')
+  strict <- isTRUE(x = getOption(x = 'Seurat.io.rds.strict', default = FALSE))
+  emit <- ifelse(test = strict, yes = abort, no = warn)
+  if (!is.null(x = cache)) {
+    if (interactive()) {
+      check_installed(pkg = 'fs', reason = 'for finding file paths')
+    } else if (!requireNamespace('fs', quietly = TRUE)) {
+      abort(message = "Loading layers from disk requires `fs`")
+    }
+    # Check the format of the cache
+    if (!is.data.frame(x = cache)) {
+      emit(message = "Malformed layer cache: not a data frame")
+      return(object)
+    }
+    if (!all(reqd.cols %in% names(x = cache))) {
+      emit(message = "Malformed layer cache: missing required columns")
+      return(object)
+    }
+    # Check the assays specified
+    assays <- .FilterObjects(object = object, classes.keep = 'StdAssay')
+    cache <- cache[cache$assay %in% assays, , drop = FALSE]
+    if (!nrow(x = cache)) {
+      emit(message = "Incorrect layer cache: none of the assays listed present")
+      return(object)
+    }
+    # Check the files
+    exists <- fs::is_file(path = cache$path)
+    exists[is.na(exists)] <- FALSE
+    cache <- cache[exists, , drop = FALSE]
+    if (!nrow(x = cache)) {
+      emit(message = "Cannot find any of the layer files specified")
+      return(object)
+    }
+    # Check the packages
+    missing.pkgs <- pkgs <- unique(x = cache$pkg)
+    for (pkg in pkgs) {
+      if (interactive()) {
+        check_installed(pkg = pkg)
+      }
+      if (requireNamespace(pkg, quietly = TRUE)) {
+        missing.pkgs <- setdiff(x = missing.pkgs, y = pkg)
+      } else {
+        emit(message = paste("Cannot find required package:", sQuote(x = pkg)))
+      }
+    }
+    pkgs <- setdiff(x = pkgs, y = missing.pkgs)
+    if (!length(x = pkgs)) {
+      emit(message = "None of the required layer packages found")
+      return(object)
+    }
+    p <- progressor(steps = nrow(x = cache))
+    # Load the layers
+    for (i in seq_len(length.out = nrow(x = cache))) {
+      lyr <- cache$layer[i]
+      pth <- cache$path[i]
+      fxn <- eval(expr = str2lang(s = cache$fxn[i]))
+      assay <- cache$assay[i]
+      p(
+        message = paste(
+          "Adding layer",
+          sQuote(x = lyr),
+          "to assay",
+          sQuote(x = assay)
+        ),
+        class = 'sticky',
+        amount = 0
+      )
+      LayerData(object = object, assay = assay, layer = lyr) <- fxn(pth)
+      p()
+    }
+  }
+  return(object)
+}
+
 #' Update old Seurat object to accommodate new features
 #'
 #' Updates Seurat objects to new structure for storing data/calculations.
@@ -1036,7 +1272,6 @@ Embeddings.Seurat5 <- function(object, reduction = 'pca', ...) {
   return(Embeddings(object = object[[reduction]], ...))
 }
 
-
 #' @param vars List of all variables to fetch, use keyword \dQuote{ident} to
 #' pull identity classes
 #' @param cells Cells to collect data for (default is all cells)
@@ -1543,6 +1778,17 @@ LayerData.Seurat <- function(object, layer = NULL, assay = NULL, ...) {
   assay <- assay %||% DefaultAssay(object = object)
   assay <- arg_match(arg = assay, values = Assays(object = object))
   return(LayerData(object = object[[assay]], layer = layer, ...))
+}
+
+#' @rdname Layers
+#' @method LayerData<- Seurat
+#' @export
+#'
+"LayerData<-.Seurat" <- function(object, layer, assay = NULL, ..., value) {
+  assay <- assay %||% DefaultAssay(object = object)
+  assay <- arg_match(arg = assay, values = Assays(object = object))
+  LayerData(object = object[[assay]], layer = layer, ...) <- value
+  return(object)
 }
 
 #' @rdname Layers
@@ -2934,6 +3180,8 @@ subset.Seurat <- function(
   }
   # Remove metadata for cells not present
   slot(object = x, name = 'meta.data') <- slot(object = x, name = 'meta.data')[cells, , drop = FALSE]
+  Idents(object = x, drop = TRUE) <- Idents(object = x)[cells]
+
   # Recalculate nCount and nFeature
   for (assay in FilterObjects(object = x, classes.keep = 'Assay')) {
     n.calc <- CalcN(object = x[[assay]])
@@ -2942,7 +3190,6 @@ subset.Seurat <- function(
       x[[names(x = n.calc)]] <- n.calc
     }
   }
-  Idents(object = x, drop = TRUE) <- Idents(object = x)[cells]
   # subset images
   for (image in Images(object = x)) {
     x[[image]] <- base::subset(x = x[[image]], cells = cells)
@@ -4265,6 +4512,8 @@ setMethod(
 #'
 #' @template return-show
 #'
+#' @rdname show-oldseurat-method
+#'
 #' @keywords internal
 #'
 #' @concept oldseurat
@@ -4605,21 +4854,20 @@ UpdateAssay <- function(old.assay, assay) {
   counts <- old.assay@raw.data
   data <- old.assay@data
   if (!inherits(x = counts, what = 'dgCMatrix')) {
-    counts <- as(object = as.matrix(x = counts), Class = 'dgCMatrix')
+    counts <- as.sparse(x = as.matrix(x = counts))
   }
   if (!is.null(x = data)) {
     if (!inherits(x = data, what = 'dgCMatrix')) {
-      data <- as(object = as.matrix(x = data), Class = 'dgCMatrix')
+      data <- as.sparse(x = as.matrix(x = data))
     }
   } else {
-    data <- as(
-      object = Matrix(
+    data <- as.sparse(
+      x = Matrix(
         data = 0,
         nrow = nrow(x = counts),
         ncol = ncol(x = counts),
         dimnames = dimnames(x = counts)
       ),
-      Class = "dgCMatrix"
     )
   }
   if (!inherits(x = old.assay@scale.data, what = 'matrix')) {